refactor(Participant): Player class 삭제 및 Participant class 변경
