+++ src/main/java/domain/Dealer.java
@@ -1,6 +1,6 @@
 package domain;
 
-public class Dealer extends Player {
+public class Dealer extends Participant {
 
     public Dealer() {
         super(new Name("딜러"));
